# swagger.yml
openapi: 3.0.0
info:
  title: BootCamp
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: user2
                password: password2
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: donkovah
                email: iamkesington@gmail.com
                password: password2
                age: 31
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...